Improve smoothness of block movement
Change the way blocks behave when being pushed so that they appear to
smoothly transition to the next tile. A conflict regarding how sounds
are fired off arises from this change; as a result, sound effects can
not be used properly for sounding off the pushing and error sounds. A
future commit might remedy this conflict.