Fix persistent solid object bug
If, upon winning, a level is played over again, there is a chance that
the last location of a given block before being moved to its goal will
persist as solid even after loading the level again. This commit fixes
that issue by ensuring that all blocks are stationary before winning.